location selection functionality added

+++ Lab2/WeatherInformer/WeatherClient/Models/CurrentWeather.cs
@@ -24,8 +24,8 @@ namespace WeatherInformer.WeatherClient.Models.CurrentWeather
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
         public double sea_level { get; set; }
@@ -35,12 +35,12 @@ namespace WeatherInformer.WeatherClient.Models.CurrentWeather
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
     {
-        public int all { get; set; }
+        public double all { get; set; }
     }
 
     public class Rain
@@ -61,8 +61,8 @@ namespace WeatherInformer.WeatherClient.Models.CurrentWeather
         public int id { get; set; }
         public double message { get; set; }
         public string country { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
     }
 
     public class CurrentWeather
@@ -71,12 +71,12 @@ namespace WeatherInformer.WeatherClient.Models.CurrentWeather
         public List<Weather> weather { get; set; }
         public string @base { get; set; }
         public Main main { get; set; }
-        public int visibility { get; set; }
+        public double visibility { get; set; }
         public Wind wind { get; set; }
         public Clouds clouds { get; set; }
         public Rain rain { get; set; }
         public Snow snow { get; set; }
-        public int dt { get; set; }
+        public double dt { get; set; }
         public Sys sys { get; set; }
         public int id { get; set; }
         public string name { get; set; }
